style and organization changes to MOAIProfiler
