+++ src/moaicore/MOAIProfilerEntry.cpp
@@ -1,0 +1,95 @@
+// Copyright (c) 2010-2011 Zipline Games, Inc. All Rights Reserved.
+// http://getmoai.com
+
+#include "pch.h"
+#include <moaicore/MOAIProfilerEntry.h>
+
+#if USE_MOAI_PROFILER
+
+#if defined ( MOAI_OS_WINDOWS )
+	#pragma optimize ( "", on ) 
+#endif
+
+//================================================================//
+// MOAIProfilerEntry
+//================================================================//
+
+MOAIProfilerEntry::MOAIProfilerEntry ( MOAIProfilerEntry* parent, const USHashedString& name ) : 
+	MOAIProfilerEntryBase ( parent, name ),
+	mRelativeDuration ( 0 ),
+	mFrameRelativeDuration ( 0 ),
+	mHitCount ( 0 ),
+	mTotalDurationMicroSec ( 0 ),
+	mMinDurationMicroSec ( 0 ),
+	mAvgDurationMicroSec ( 0 ),
+	mMaxDurationMicroSec ( 0 ),
+	mUnusedFrames ( 0 ) {
+
+	memset ( mTotalDurationSamplesMicroSec, 0, sizeof( mTotalDurationSamplesMicroSec ) );
+}
+	
+//----------------------------------------------------------------//
+void MOAIProfilerEntry::Reset () {
+
+	MOAIProfilerEntryBase::Reset ();
+
+	mRelativeDuration = 0;
+	mFrameRelativeDuration = 0;
+
+	mHitCount = 0;
+
+	mTotalDurationMicroSec = 0;
+	memset ( mTotalDurationSamplesMicroSec, 0, sizeof( mTotalDurationSamplesMicroSec ) );
+
+	mMinDurationMicroSec = UINT_MAX;
+	mAvgDurationMicroSec = 0;
+	mMaxDurationMicroSec = 0;
+
+	mUnusedFrames = 0;
+}
+
+//----------------------------------------------------------------//
+bool MOAIProfilerEntry::IsLessThan ( const MOAIProfilerEntryBase& otherBase ) const {
+
+	const MOAIProfilerEntry& other = (const MOAIProfilerEntry&)otherBase;
+
+	return mAvgDurationMicroSec < other.mAvgDurationMicroSec;
+}
+
+//----------------------------------------------------------------//
+void MOAIProfilerEntry::CopyFrom ( const MOAIProfilerEntryBase& otherBase ) {
+
+	const MOAIProfilerEntry& other = (const MOAIProfilerEntry&)otherBase;
+
+	mRelativeDuration = other.mRelativeDuration;
+	mFrameRelativeDuration = other.mFrameRelativeDuration;
+
+	mHitCount = other.mHitCount;
+
+	mTotalDurationMicroSec = other.mTotalDurationMicroSec;
+	memcpy ( mTotalDurationSamplesMicroSec, other.mTotalDurationSamplesMicroSec, sizeof( mTotalDurationSamplesMicroSec ) );
+
+	mMinDurationMicroSec = other.mMinDurationMicroSec;
+	mAvgDurationMicroSec = other.mAvgDurationMicroSec;
+	mMaxDurationMicroSec = other.mMaxDurationMicroSec;
+
+	mUnusedFrames = other.mUnusedFrames;
+}
+
+//----------------------------------------------------------------//
+void MOAIProfilerEntry::SwapWith ( MOAIProfilerEntryBase& otherBase ) {
+	
+	MOAIProfilerEntry& other = (MOAIProfilerEntry&)otherBase;
+
+	static MOAIProfilerEntry swap ( 0, USHashedString::Empty );
+
+	swap.CopyFrom ( other );
+	other.CopyFrom ( *this );
+	CopyFrom ( swap );
+}
+
+#if defined ( MOAI_OS_WINDOWS )
+	#pragma optimize ( "", off ) 
+#endif
+
+#endif
