+++ src/moaicore/MOAIProfilerEntryBase.h
@@ -1,0 +1,45 @@
+// Copyright (c) 2010-2011 Zipline Games, Inc. All Rights Reserved.
+// http://getmoai.com
+
+#ifndef	MOAIPROFILERENTRYBASE_H
+#define	MOAIPROFILERENTRYBASE_H
+
+#if USE_MOAI_PROFILER
+
+//================================================================//
+// MOAIProfilerEntryBase
+//================================================================//
+class MOAIProfilerEntryBase {
+protected:
+
+	//----------------------------------------------------------------//
+	virtual void			CopyFrom				( const MOAIProfilerEntryBase& other );
+							MOAIProfilerEntryBase	( MOAIProfilerEntryBase* parent, const USHashedString& name );
+	virtual void			SwapWith				( MOAIProfilerEntryBase& other ) = 0;
+
+private:
+	
+	//----------------------------------------------------------------//
+	void						DebugChildren	( std::string& childNames ) const;
+	static void					SwapChildren	( MOAIProfilerEntryBase& left, MOAIProfilerEntryBase& right );
+	
+public:
+
+	MOAIProfilerEntryBase*		mNext;
+	USHashedString				mName;
+	u16							mDepth;
+	MOAIProfilerEntryBase*		mParent;
+	MOAIProfilerEntryBase*		mFirstChild;
+	MOAIProfilerEntryBase*		mLastChild;
+	u32							mNumChildren;
+	
+	//----------------------------------------------------------------//
+	void						AddChildEntry			( MOAIProfilerEntryBase* child );
+	virtual bool				IsLessThan				( const MOAIProfilerEntryBase& other ) const = 0;
+	virtual						~MOAIProfilerEntryBase	() {}
+	virtual void				Reset					();
+	void						SortChildren			();
+};
+
+#endif
+#endif
