+++ src/moaicore/MOAIProfilerScope.h
@@ -1,0 +1,37 @@
+// Copyright (c) 2010-2011 Zipline Games, Inc. All Rights Reserved.
+// http://getmoai.com
+
+#ifndef	MOAIPROFILERSCOPE_H
+#define	MOAIPROFILERSCOPE_H
+
+#include <moaicore/MOAIMutex.h>
+#include <moaicore/MOAIProfilerReport.h>
+#include <moaicore/MOAIProfilerAutoScope.h>
+
+#if USE_MOAI_PROFILER
+
+//================================================================//
+// MOAIProfilerScope
+//================================================================//
+class MOAIProfilerScope {
+private:
+	
+	friend class MOAIProfilerContext;
+	
+	//----------------------------------------------------------------//
+	MOAIProfilerScope*				mNext;
+	MOAIProfilerScope*				mParent;
+	USHashedString					mName;
+	USDeviceTime::TimeStamp			mDuration;
+
+public:
+
+	//----------------------------------------------------------------//
+	void		Enter					( MOAIProfilerScope* parent, const USHashedString& name );
+	void		Leave					( const USHashedString& name );
+	void		Reset					();
+				MOAIProfilerScope		( const USHashedString& name = USHashedString::Empty );
+};
+
+#endif
+#endif
